requirements.txt
attrs==25.1.0
boule==0.5.0
contourpy==1.2.1
cycler==0.12.1
fonttools==4.55.8
gtsam==4.2
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.8.0
narwhals==1.29.1
numpy==1.22.3
packaging==24.2
pandas==1.5.3
pillow==11.1.0
plotly==6.0.0
pyparsing==3.2.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
scipy==1.11.4
six==1.17.0
tqdm==4.67.1
zipp==3.21.0
